Calculator style tweaks
